clean up citysavedlist query renderer
