+++ src/lib/Containers/index.tsx
@@ -1,5 +1,4 @@
 import { ArtworkContainer as Artwork } from "lib/Scenes/Artwork"
-import CitySavedList from "lib/Scenes/City/CitySavedList"
 import CitySectionList from "lib/Scenes/City/CitySectionList"
 import { PartnerContainer as Partner } from "lib/Scenes/Partner"
 import Fair from "../Scenes/Fair"
@@ -9,7 +8,6 @@ import Show from "../Scenes/Show"
 
 export default {
   Artwork,
-  CitySavedList,
   CitySectionList,
   Fair,
   Home,
